requirements.txt
imageio==2.6.1 # only used in space_vis.py
matplotlib~=3.5.0
numpy~=1.21.4
Pillow~=9.5.0 #must be lower than 10.0.0 to still work with ANTIALIAS
protobuf~=3.20.0
scikit-image>=0.19.0
scipy~=1.7.3
tqdm>=4.64.1
yacs~=0.1.8
tensorboard>=2.9.1
rtpt~=0.0.4
torch>=1.12.0
opencv-python>=4.5.5
torchvision>=0.13.0
pandas>=1.3.0
scikit-learn~=1.0.1
termcolor~=1.1.0 # not used that much, but only used for printing in the terminal
NicePrinter
ocatari ~= 2.1.0
gymnasium[atari, accept-rom-license]
pyclustering